refactor: renaming providers
