+++ src/providers/openrank/types.ts
@@ -1,3 +1,18 @@
+export interface TopProfile {
+    fid: number;
+    fname: string;
+    username: string;
+    rank: number;
+    score: number;
+    percentile: number;
+}
+
+export interface PostForYouByAuthorship {
+    cast_hash: string;
+    fid: number;
+    timestamp: string;
+}
+
 export interface FeedPost {
     postId: string;
     handle: string;
